% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability_v3.R
\name{fun_Cp}
\alias{fun_Cp}
\title{Calculate index for process capability, i.e. Cp, Cpk, Pp, Ppk...}
\usage{
fun_Cp(subgroup = 1, x, USL = NULL, target = NULL, LSL = NULL)
}
\arguments{
\item{subgroup}{subgroup No. or Id}

\item{x}{numeric vector for calculating process capability index}

\item{USL}{upper specification limit}

\item{LSL}{lower specification limit}

\item{targt}{spec target}
}
\description{
Calculate index for process capability, i.e. Cp, Cpk, Pp, Ppk...
}
