% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability.R
\name{fun_proCap_binom_plot}
\alias{fun_proCap_binom_plot}
\title{Binomial process capability analysis}
\usage{
fun_proCap_binom_plot(size, data, USL = NA, target = NA, LSL = NA,
  df_info = NULL, info_names = NULL, bar_fill = "#36648B")
}
\arguments{
\item{size}{numeric vector of inspection size}

\item{data}{numeric vector with same length as size, events or defects counts}

\item{USL}{upper specification limit of process}

\item{target}{target (center) of process}

\item{LSL}{lower specification limit of process}

\item{df_info}{dataframe of information shown on hover}

\item{info_names}{character vector with the same length as data,
to replace names(info) to show on hover. It doesn't include x-variable and
y-varialbe. xlab and ylab will be used for x-variable and y-variable.}

\item{bar_fill}{bar filling color}
}
\value{
list contains sample statistics, QQ plot and histogram
}
\description{
Binomial process capability analysis
}
