% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability.R
\name{fun_proCap_norm_plot}
\alias{fun_proCap_norm_plot}
\title{Normal process capability analysis}
\usage{
fun_proCap_norm_plot(data, group = 1, USL = NA, target = NA,
  LSL = NA, df_info = NULL, info_names = NULL,
  bar_fill = "#36648B")
}
\arguments{
\item{data}{numeric vector}

\item{group}{vector with the same length of data to form group for within and
between group}

\item{USL}{upper specification limit of process}

\item{target}{target (center) of process}

\item{LSL}{lower specification limit of process}

\item{df_info}{dataframe of information shown on hover}

\item{info_names}{character vector with the same length as data,
to replace names(info) to show on hover. It doesn't include x-variable and
y-varialbe. xlab and ylab will be used for x-variable and y-variable.}

\item{bar_fill}{bar filling color}
}
\value{
list contains sample statistics, QQ plot and histogram
}
\description{
Normal process capability analysis
}
