% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability_v3.R
\name{fun_proCap_plot}
\alias{fun_proCap_plot}
\title{Normal process capability analysis}
\usage{
fun_proCap_plot(subgroup = 1, x, USL = NA, target = NA, LSL = NA)
}
\description{
Normal process capability analysis
}
