% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability_v3.R
\name{fun_proCap_pois_plot}
\alias{fun_proCap_pois_plot}
\title{Poisson process capability analysis}
\usage{
fun_proCap_pois_plot(size, x, iu, USL = NA, target = NA, LSL = NA)
}
\arguments{
\item{size}{vector of (inspection) size}

\item{x}{vector of data}

\item{iu}{inspection unit, how many samples inspected per unit}
}
\description{
Poisson process capability analysis
}
