% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability_v3.R
\name{fun_qq_plot}
\alias{fun_qq_plot}
\title{Plot Q-Q plot by ggplot}
\usage{
fun_qq_plot(x, title = "Normal Q-Q Plot", xlab = "Theoretical",
  ylab = "Sample")
}
\arguments{
\item{x}{numeric vector}

\item{xlab}{x-axis label}
}
\description{
Plot Q-Q plot by ggplot
}
