% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability.R
\name{Plot_hist_norm}
\alias{Plot_hist_norm}
\title{Plot histogram for process capability}
\usage{
Plot_hist_norm(data, mean = NULL, StDev_overall = NULL,
  StDev_within = NULL, plot_spec = NULL,
  title = "Sample Distribution", xlab = "", ylab = "",
  bar_fill = "#36648B")
}
\arguments{
\item{data}{vector of data to plot histogram}

\item{mean}{numeric, if NUll will be count by data}

\item{StDev_overall}{numeric, if NUll will be count by data}

\item{StDev_within}{numeric, if NUll will be count by data}

\item{plot_spec}{vector with three elements, c(USL= 1, target= 0, LSL= -1)}

\item{title}{plot title}

\item{xlab}{x label}

\item{ylab}{y label}

\item{bar_fill}{bar filling color}
}
\value{
histogram
}
\description{
Plot histogram for process capability
}
