% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability_v3.R
\name{Plotly_qq}
\alias{Plotly_qq}
\title{Plot Q-Q plot by plotly}
\usage{
Plotly_qq(x, info = NULL, infolab = NULL, title = "Normal Q-Q Plot",
  xlab = "Theoretical", ylab = "Sample")
}
\arguments{
\item{x}{vector of data to plot QQ}

\item{info}{dataframe of information shown on hover}

\item{infolab}{character vector to replace names(info) to show on hover}

\item{title}{plot title}
}
\description{
Plot Q-Q plot by plotly
}
