% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability.R
\name{QQ_line}
\alias{QQ_line}
\title{return a dataframe with x, y coordinates for plotting QQ line}
\usage{
QQ_line(data, distribution = "norm", dparams = list(),
  qprobs = c(0.25, 0.75), detrend, identity, qtype)
}
\description{
return a dataframe with x, y coordinates for plotting QQ line
}
