% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability.R
\name{QQ_plot}
\alias{QQ_plot}
\title{Plot Q-Q plot by plotly}
\usage{
QQ_plot(data, group = 1, dist = "norm", dparams = list(),
  df_info = NULL, info_names = NULL, title = "Q-Q Plot",
  xlab = "Theoretical", ylab = "Sample", qprobs = c(0.25, 0.75),
  detrend = FALSE, identity = FALSE, qtype = 7,
  colors = color_set4, plotly = TRUE)
}
\arguments{
\item{data}{vector of data to plot QQ}

\item{group}{vector with the same length as data, to group data}

\item{dist}{character, theorectical distribution function to use,
e.g. "norm", "binom", "pois"... (refer more to package 'qqplotr')}

\item{dparams}{list of additional parameters passed on to the previous chosen
distribution funcion.}

\item{df_info}{dataframe of information shown on hover}

\item{info_names}{character vector with the same length as data,
to replace names(info) to show on hover. It doesn't include x-variable and
y-varialbe. xlab and ylab will be used for x-variable and y-variable.}

\item{title}{plot title}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{qprobs}{numeric vector with length 2. Rrepresents the quantitles used to
construct the Q-Q line}

\item{detrend}{logic, should the plot objexts be detrended?}

\item{identity}{logic, should an identity line be used as the reference line used to
construct the confidence bands?}

\item{qtype}{interger between 1 and 9, Type of quantile algorithm to be used by
the quantile function to construct the Q-Q line.}

\item{colors}{colors set}

\item{plotly}{logic, TRUE for plotly, FALSE for ggplot}
}
\description{
Plot Q-Q plot by plotly
}
