% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability.R
\name{fun_Cp}
\alias{fun_Cp}
\title{Calculate index for process capability, i.e. Cp, Cpk, Pp, Ppk...}
\usage{
fun_Cp(data, group = 1, USL = NULL, target = NULL, LSL = NULL)
}
\arguments{
\item{data}{numeric vector for calculating process capability index}

\item{group}{vector with the same length of data to form group for within and
between group}

\item{USL}{upper specification limit}

\item{LSL}{lower specification limit}

\item{targt}{spec target}
}
\value{
list contains: sample_size, sample_mean, StDev_overall,
        StDev_within (within-group), Cp, Cpk, CpkU, CpkL, Pp, Ppk, PpkU, PpkL,
        ppm_within, ppm_overall, ppm_obs
}
\description{
Calculate index for process capability, i.e. Cp, Cpk, Pp, Ppk...
}
