% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability_v3.R
\name{fun_proCap_binom_plot}
\alias{fun_proCap_binom_plot}
\title{Binomial process capability analysis}
\usage{
fun_proCap_binom_plot(size, x, USL = NA, target = NA, LSL = NA)
}
\arguments{
\item{size}{vector of (inspection) size}

\item{x}{vector of data}
}
\description{
Binomial process capability analysis
}
