% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCapability_v3.R
\name{Plot_hist_norm}
\alias{Plot_hist_norm}
\title{Plot histogram for process capability}
\usage{
Plot_hist_norm(x, mean = NULL, StDev_overall = NULL,
  StDev_within = NULL, plot_spec = NULL,
  title = "Sample Distribution")
}
\arguments{
\item{x}{vector of data to plot histogram}

\item{plot_spec}{c(USL= 1, target= 0, LSL= -1)}

\item{title}{plot title}
}
\description{
Plot histogram for process capability
}
